4DEnVAR en Fortran

+++ FORTRAN/Lorenz_EnKF/functions/Lorenz_96_one_Step.F90
@@ -1,9 +1,9 @@
   subroutine Lorenz_96_One_Step(n,dt,x0,F,X)
     integer,intent(in)::n
-    real,intent(in)::dt,F
-    real,dimension(n),intent(in)::x0
-    real,dimension(n),intent(out)::X
-    real,dimension(n)::dx
+    real*8,intent(in)::dt,F
+    real*8,dimension(n),intent(in)::x0
+    real*8,dimension(n),intent(out)::X
+    real*8,dimension(n)::dx
 
     dx(1) = (x0(2)-x0(n-1))*x0(n)-x0(1)+F
     dx(2) = (x0(3)-x0(n))*x0(1)-x0(2)+F
